Fixed the front page view!

+++ checkout/views.py
@@ -29,7 +29,10 @@ def payment_info(request):
             user_payment_info = form.save(commit=False)
             user_payment_info.user = request.user
             user_payment_info.save()
-            return redirect("review_order")
+            if "contact_info" in request.POST:
+                return redirect("contact_info")
+            elif "review_order" in request.POST:
+                return redirect("review_order")
     else:
         form = PaymentDetailsForm(instance=user_payment_info)
     context = {"form": form}
@@ -47,11 +50,13 @@ def review_step(request):
 
 
 def confirmation(request):
-    user_cart = Cart.objects.filter(user=request.user).first()
-    user_payment_info = PaymentDetails.objects.filter(user=request.user).first()  # needs to be erased when POST
     user_contact_info = ContactInfo.objects.filter(user=request.user).first()  # needs to be erased when POST
-    user_cart.clear_cart()
-    user_cart.save()
-    user_payment_info.delete()
-    user_contact_info.delete()
+    if user_contact_info:
+        user_contact_info.delete()
+        user_cart = Cart.objects.filter(user=request.user).first()
+        user_cart.clear_cart()
+        user_cart.save()
+        user_payment_info = PaymentDetails.objects.filter(user=request.user).first()  # needs to be erased when POST
+        if user_payment_info:
+            user_payment_info.delete()
     return render(request, "checkout/thank_you.html")
